app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
firebase-admin==6.3.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
cloud-sql-python-connector==1.11.0
pytest==7.4.3
requests==2.31.0
